Handle tcp product

+++ Server/Server/SQL/EmployeeHandler.cs
@@ -1,5 +1,6 @@
 using System;
 using MySql.Data.MySqlClient;
+using Org.BouncyCastle.Asn1.Cms;
 using Renci.SshNet.Messages.Connection;
 
 namespace Server.SQL
@@ -8,13 +9,13 @@ namespace Server.SQL
     {
         private const string SqlLogin = @"server=localhost;userid="+SQLCredentials.MySQLUsername+";password="+SQLCredentials.MySQLPassword+";database=myhome";
 
-        public Employee GetEmployeeById(int id)
+        public String GetEmployeeById(int id)
         {
             try
             {
                 using var con = new MySqlConnection(SqlLogin);
                 con.Open();
-
+                string result = "";
                 string query = "SELECT * FROM users WHERE id = " + id;
                 using var cmd = new MySqlCommand(query, con);
                 Employee employee = null;
@@ -26,35 +27,55 @@ namespace Server.SQL
                     //Console.WriteLine("{0} {1} {2} {3}", reader.GetInt32(0), reader.GetString(1), reader.GetString(2), reader.GetString(3));
                 }
 
-                return employee;
+                if (employee == null)
+                    return "ERROR: Employee doesn't exist in the database.";
+
+                result = employee.employeeID + " ! " + employee.employeeName + " ! " + employee.role + " ! " +
+                         employee.password;
+
+                return result;
             }
             catch (Exception e)
             {
                 Console.WriteLine(e);
-                return null;
+                return "Error occoured.";
             }
         }
 
-        public void DeleteEmployeeFromDB(int id)
+        public string DeleteEmployeeFromDB(int id)
         {
             try
             {
                 using var con = new MySqlConnection(SqlLogin);
                 con.Open();
-                string query = "DELETE FROM users where id = " + id;
+                string query = "SELECT * FROM users WHERE id = " + id;
+                using var command = new MySqlCommand(query, con);
+                using MySqlDataReader reader = command.ExecuteReader();
+                int DbID = 0;
+                while (reader.Read())
+                {
+                    DbID = reader.GetInt32(0);
+                }
+                reader.Close();
+
+                if (DbID == 0)
+                {
+                    return "ERROR: Employee doesn't exist in the database.";
+                }
+                query = "DELETE FROM users where id = " + id;
                 using var cmd = new MySqlCommand(query, con);
                 cmd.ExecuteNonQuery();
                 con.Close();
-                Console.WriteLine("Removed user associated with id: " + id);
+                return "Removed user associated with id: " + id;
             }
             catch (Exception e)
             {
                 Console.WriteLine(e);
-                throw;
+                return "ERROR: Employee doesn't exist in the database.";
             }
         }
 
-        public void AddUserToDB(int id, string pw, string name, Employee.Role role)
+        public string AddUserToDB(int id, string pw, string name, string role)
         {
             try
             {
@@ -74,9 +95,8 @@ namespace Server.SQL
 
                  if (ids != 0)
                  {
-                     Console.WriteLine("User already associated with this id.");
                      con.Close();
-                     return;
+                     return "ERROR: Employee already exists.";
                  }
 
                  var sql = "INSERT INTO users(id, username, password, role, register_date) values (@id, @username, @password, @role, @register_date)";
@@ -86,18 +106,19 @@ namespace Server.SQL
                  cmd.Parameters.AddWithValue("@id", id);
                  cmd.Parameters.AddWithValue("@username", name);
                  cmd.Parameters.AddWithValue("@password", pw);
-                 cmd.Parameters.AddWithValue("@role", Enum.GetName(typeof(Employee.Role),role));
+                 //TODO: Make sure the worker role only can be "OfficeWorker" or "FloorWorker"
+                 cmd.Parameters.AddWithValue("@role", role);
                  cmd.Parameters.AddWithValue("@register_date", DateTime.Now);
                  cmd.Prepare();
          
                  cmd.ExecuteNonQuery();
-       
-                 Console.WriteLine("Added user!");
+
+                 return "Added employee to the database.";
             }
             catch (Exception e)
             {
                  Console.WriteLine(e);
-                 throw;
+                 return "ERROR: Employee already exists.";
             }
         }
 
@@ -119,7 +140,7 @@ namespace Server.SQL
             }
         }
 
-        public bool AuthenticateUser(int id, string pw)
+        public string AuthenticateUser(int id, string pw)
         {
             try
             {
@@ -131,9 +152,12 @@ namespace Server.SQL
                 while (reader.Read())
                 {
                     if (reader.GetInt32(0) == id && reader.GetString(2).Equals(pw))
-                        return true;
+                    {
+                        return reader.GetInt32(0) + " ! " + reader.GetString(1) + " ! " + reader.GetString(2) + " ! " +
+                               reader.GetString(3);
+                    }
                 }
-                return false;
+                return "ERROR: Incorrect information.";
             }
             catch (Exception e)
             {
